requirements.txt
amqp==2.3.2
asgiref==3.2.3
asn1crypto==0.24.0
async-generator==1.10
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
beautifulsoup4==4.7.1
billiard==3.5.0.5
bs4==0.0.1
celery==4.2.1
celery-once==2.0.0
certifi==2018.11.29
cffi==1.12.3
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.0
Django==3.0rc1
django-bootstrap4==1.0.1
django-crispy-forms==1.8.1
django-djconfig==0.9.0
django-environ==0.4.5
django-extensions==2.1.4
django-jet==1.0.8
django-jsoneditor==0.1.4
django-redis-cache==2.1.0
django-timezone-field==3.0
django-yamlfield==1.0.3
enum34==1.1.6
feedparser==5.2.1
future==0.17.1
h2==3.1.1
hiredis==1.0.1
hpack==3.0.0
humanize==0.5.1
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
json-rpc==1.12.1
jsonfield2==3.0.3
jsonschema==3.2.0
kombu==4.2.2.post1
markovify==0.7.2
more-itertools==7.2.0
oauthlib==3.0.1
packaging==19.2
Pillow==5.4.1
priority==1.3.0
psycopg2-binary==2.7.6.1
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
pyparsing==2.4.5
Pyrogram==0.11.0
pyrsistent==0.15.6
PySocks==1.6.8
pyTelegramBotAPI==3.6.6
python-anticaptcha==0.3.1
python-crontab==2.3.6
python-dateutil==2.7.5
python-telegram-bot==12.2.0
python-twitter==3.5
pytz==2018.9
PyYAML==3.13
redis==3.0.1
requests==2.21.0
requests-oauthlib==1.2.0
sentry-sdk==0.6.9
service-identity==18.1.0
six==1.12.0
soupsieve==1.7.3
sqlparse==0.3.0
TgCrypto==1.1.1
tornado==6.0.3
Twisted==19.10.0
txaio==18.8.1
Unidecode==1.1.1
urllib3==1.24.1
vine==1.2.0
vk-api==11.6.1
Werkzeug==0.14.1
zipp==0.6.0
zope.interface==4.7.1
